+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+-Command Queue 0.0.3 (2018-09-16)
+--------------------------------
+- Fix possible edge case of a receiver setting a different receiver in the `receiveCommand` block.
+
+- Add sample code.
+
 -Command Queue 0.0.2 (2018-09-16)
 --------------------------------
 - Add `hasReceiver()`.